only delete if log file exists

+++ FileStreamRotator.js
@@ -255,7 +255,9 @@ FileStreamRotator.writeAuditLog = function(audit){
 function removeFile(file){
     if(file.hash === crypto.createHash('md5').update(file.name + "LOG_FILE" + file.date).digest("hex")){
         try{
-            fs.unlinkSync(file.name);
+            if (fs.existsSync(file.name)) {
+                fs.unlinkSync(file.name);
+            }
         }catch(e){
             console.error(new Date(), "[FileStreamRotator] Could not remove old log file: ", file.name);
         }
